added student card maker module
