+++ index.html
@@ -1,15 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
-  <link rel="stylesheet" href="main.css">
-  <title>Sorting Hats || Find Your School</title>
-</head>
-
-<body>
-  
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
+    <link rel="stylesheet" href="main.css">
+    <title>Sorting Hats || Find Your School</title>
+  </head>
+  <body>
     <div class="jumbotron d-flex flex-wrap justify-content-around bg-info bg-gradient">
       <h1 class="display-4">Welcome to Hogwarts!</h1>
       <p class="lead">Welcome to Hogwarts! Find which school you are by wearing the sorting hat. Click below to begin you in-class experience.</p>
@@ -19,16 +17,15 @@
         <a id="sort" type="submit" class="btn btn-primary btn-lg" href="#" role="button">Sort!</a>
       </p>
     </div>
-    
-    <div id="form"></div>
-    
-    <div id="students" class="d-flex flex-wrap justify-content-around"></div> 
       
-
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js" 
-          integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW" 
-          crossorigin="anonymous">
-  </script>
-  <script src="main.js" type="module"></script>
-</body>
+    <div id="form"></div>
+      
+    <div id="students" class="d-flex flex-wrap justify-content-around"></div> 
+        
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js" 
+            integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW" 
+            crossorigin="anonymous">
+    </script>
+    <script src="./scripts/main.js" type="module"></script>
+  </body>
 </html>
