+++ scripts/main.js
@@ -1,4 +1,4 @@
-import printToDom  from './printToDom.js';
+import studentMaker from './studentMaker.js'
 
 const students = [];
 
@@ -8,23 +8,6 @@ const houses = [
   "Ravenclaw",
   "Slytherin",
 ]
-
-// Prints student card
-const studentMaker = (studentArr) => {
-  let cardString = '';
-
-  studentArr.forEach((item, i) => {
-    cardString += `<div class="card" style="width: 18rem;" id="${i}">
-                    <div class="card-body ${item.style}">
-                      <h5 class="card-title text-light">${item.name}</h5>
-                      <p class="card-text text-light">${item.school}</p>
-                      <button type="button" class="btn btn-danger" id=${item.id}>Expel</button>
-                    </div>
-                  </div>`;
-  });
-
-  printToDom("#students", cardString);
-}
 
 // Creates student object to be pushed into "student" array with input from form id
 const createStudentObj = (e) => {
